backend-develop/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.5.1
Deprecated==1.2.13
Django==3.2.9
django-dotenv==1.4.2
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
pymongo==4.0.1
pyparsing==3.0.6
psycopg2
pytz==2021.3
redis==4.0.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
zipp==3.6.0
